login_service/requirements.txt
pact-python
